Split PackageManager's Program.cs in multiple files
